GCC compile fixes

+++ BlockSettleUILib/Trading/ReqCCSettlementContainer.cpp
@@ -62,7 +62,7 @@ bs::sync::PasswordDialogData ReqCCSettlementContainer::toPasswordDialogData() co
    bs::sync::PasswordDialogData dialogData = SettlementContainer::toPasswordDialogData();
    dialogData.setValue(keys::Market, "CC");
    dialogData.setValue(keys::AutoSignCategory, static_cast<int>(bs::signer::AutoSignCategory::SettlementRequestor));
-   dialogData.setValue(keys::LotSize, lotSize_);
+   dialogData.setValue(keys::LotSize, qint64(lotSize_));
 
    dialogData.remove(keys::SettlementId);
